Move IJEPAMaskCollator out of lightly.data.collate
IJEPAMaskCollator is not deprecated but sits in collate.py next to the
collate functions that are, so it would go down with them. Give it its own
module and export it from lightly.data.

Body is unchanged apart from type annotations, which the new module needs
since collate.py is on the mypy exclude list and the new one isn't.
